requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.2
xgboost==2.1.4
scikit-learn==1.7.2
shap==0.47.2
pandas==3.0.1
numpy==2.4.3
supabase==2.4.6
python-dotenv==1.0.1
pydantic==2.12.5
